make DynamicConfigProvider generic
